+++ lib/src/pages/home_page.dart
@@ -1,7 +1,7 @@
 import 'package:briefcase/src/pages/dashboard_page.dart';
 import 'package:briefcase/src/pages/history_page.dart';
 import 'package:briefcase/src/pages/login_page.dart';
-import 'package:briefcase/src/pages/transaction_page.dart';
+import 'package:briefcase/src/pages/transactions/transaction_page.dart';
 import 'package:flutter/material.dart';
 
 class HomePage extends StatefulWidget {
