+++ lib/src/pages/transactions/transactionsHistory_page.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class TransactionHistoryPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(child: Text('historial'));
+  }
+}
